Minor Improvements
Added a README in /Scripts with basic directions.
Added my local bash.bashrc as a backup.
Added builddc.bat and rundc.bat for consistency with Connected Farm
scripts.
Created an AutoIT script because BitFender was blocking SendKeys from
cmd.
The runcf script will now open a new Chrome window with the CF url and
open dev tools. The amount of time it took to automatically press "F12"
was ridiculous, but I learned a little AutoIT.